+++ micro-ui/web/micro-ui-internals/packages/modules/core/src/components/RoleBasedEmployeeHome.js
@@ -1,7 +1,8 @@
 import React, { Fragment, useState, useEffect } from "react";
-import { Button, LandingPageCard, LandingPageWrapper, Loader } from "@egovernments/digit-ui-components";
+import { Button, LandingPageCard, LandingPageWrapper } from "@egovernments/digit-ui-components";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
+import { Loader } from "@egovernments/digit-ui-components";
 
 
 
@@ -116,7 +117,7 @@ export const RoleBasedEmployeeHome = ({ modules, additionalComponent }) => {
     }, {});
 
   if (isLoading) {
-    return <Loader />;
+    return <Loader page={false} variant={"PageLoader"} />;
   }
 
   if (!configEmployeeSideBar) {