+++ micro-ui/web/micro-ui-internals/packages/modules/workbench/src/components/WorkbenchCard.js
@@ -26,11 +26,11 @@ const WorkbenchCard = () => {
       link: `/${window?.contextPath}/employee/workbench/localisation-search`,
       roles: ROLES.LOCALISATION,
     },
-    {
-      label: t("ACTION_TEST_UPLOAD_BOUNDARY"),
-      link: `/${window?.contextPath}/employee/workbench/upload-boundary`,
-      roles: ROLES.LOCALISATION,
-    },
+    // {
+    //   label: t("ACTION_TEST_UPLOAD_BOUNDARY"),
+    //   link: `/${window?.contextPath}/employee/workbench/upload-boundary`,
+    //   roles: ROLES.LOCALISATION,
+    // },
     // {
     //   label: t("Sample Create master"),
     //   link: `/${window?.contextPath}/employee/workbench/mdms-add-v2?moduleName=common-masters&masterName=Sample`,