+++ micro-ui/web/micro-ui-internals/packages/modules/core/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog 
 
+## [1.8.22]  [3-Feb-2025]
+- FEATURE/HCMPRE-2208  : Fixed some loader component issue 
+- added new prop called allowedUserTypes to restrict any user type if needed
 
 ## [1.8.21]  [1-Feb-2025]
 - FEATURE/HCMPRE-1425 : Added the workbench module patches and Updated localisation search screen, and core module #2181